backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
qdrant-client==1.7.0
openai==1.10.0
boto3==1.34.34
pypdf==3.17.4
orjson==3.9.12

# OCR support (optional but recommended)
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
